requirements.txt
pandas~=0.24.2
scikit-learn~=0.21.2
tensorflow~=1.13.1
numpy~=1.16.4
setuptools~=41.0.1
python-dotenv~=0.14.0
xgboost~=0.90
